Rename lint to needless_take_by_value
And fixes false-positives for generics and `match`
