+++ tests/ui/absurd-extreme-comparisons.rs
@@ -2,7 +2,7 @@
 #![plugin(clippy)]
 
 #![deny(absurd_extreme_comparisons)]
-#![allow(unused, eq_op, no_effect, unnecessary_operation, needless_take_by_value)]
+#![allow(unused, eq_op, no_effect, unnecessary_operation, needless_pass_by_value)]
 
 fn main() {
     const Z: u32 = 0;
