+++ tests/ui/should_assert_eq.rs
@@ -1,7 +1,7 @@
 #![feature(plugin)]
 #![plugin(clippy)]
 
-#![allow(needless_take_by_value)]
+#![allow(needless_pass_by_value)]
 #![deny(should_assert_eq)]
 
 #[derive(PartialEq, Eq)]
